improve macro/bridge api

+++ src/macro/sim/types.hpp
@@ -10,30 +10,30 @@
 namespace mind_sim::macro::sim {
 
 struct ScalarBuffer {
-    std::vector<float> values{};
+    std::vector<double> values{};
 
     ScalarBuffer() = default;
-    explicit ScalarBuffer(std::size_t exposure_count): values(exposure_count, 0.0F) {}
+    explicit ScalarBuffer(std::size_t exposure_count): values(exposure_count, 0.0) {}
 
     [[nodiscard]] std::size_t size() const noexcept {
         return values.size();
     }
 
-    [[nodiscard]] float get(int exposure_id, std::string_view context) const {
+    [[nodiscard]] double get(int exposure_id, std::string_view context) const {
         if (exposure_id < 0 || static_cast<std::size_t>(exposure_id) >= values.size()) {
             throw std::runtime_error(std::string(context) + " exposure index out of range");
         }
         return values[static_cast<std::size_t>(exposure_id)];
     }
 
-    [[nodiscard]] float& at(int exposure_id, std::string_view context) {
+    [[nodiscard]] double& at(int exposure_id, std::string_view context) {
         if (exposure_id < 0 || static_cast<std::size_t>(exposure_id) >= values.size()) {
             throw std::runtime_error(std::string(context) + " exposure index out of range");
         }
         return values[static_cast<std::size_t>(exposure_id)];
     }
 
-    void fill(float value) {
+    void fill(double value) {
         std::fill(values.begin(), values.end(), value);
     }
 };
@@ -42,7 +42,7 @@ struct ExposureRecord {
     int roi_count{0};
     int exposure_count{0};
     std::vector<int> roi_indices{};
-    std::vector<float> values{};
+    std::vector<double> values{};
 
     [[nodiscard]] std::size_t recorded_roi_count() const noexcept {
         return roi_indices.size();
